% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomFuns.R
\name{bestLoess}
\alias{bestLoess}
\title{bestLoess}
\usage{
bestLoess(model, criterion = c("aicc", "aicc1", "gcv"), spans = c(0.05,
  0.95))
}
\arguments{
\item{model}{}

\item{criterion}{}

\item{spans}{}
}
\description{
bestLoess
}

