% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{classify.psa.raster}
\alias{classify.psa.raster}
\title{classify.psa.raster}
\usage{
classify.psa.raster(MUD, SAND, GRAV, max = 1)
}
\arguments{
\item{MUD}{}

\item{SAND}{}

\item{GRAV}{}

\item{max}{}
}
\description{
classify.psa.raster
}

