% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{classify.psa.vector}
\alias{classify.psa.vector}
\title{classify.psa.vector}
\usage{
classify.psa.vector(MUD, SAND, GRAV, max = 100)
}
\arguments{
\item{MUD}{}

\item{SAND}{}

\item{GRAV}{}

\item{max}{}
}
\description{
classify.psa.vector
}

