% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coast}
\alias{coast}
\title{DATA - Vector coastline of europe}
\format{A data frame with 50 rows:
\describe{
  \item{NATION}{ - }
  \item{CNTRYNAME}{name}
  \item{CNTRYABBR}{abr name}
  ...
}}
\usage{
coast
}
\description{
Source is ESRI I believe
}
\keyword{datasets}

