% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomFuns.R
\name{didYouMean}
\alias{didYouMean}
\title{didYouMean}
\usage{
didYouMean(input)
}
\arguments{
\item{input}{string}
}
\value{
sting
}
\description{
returns google suggestion for a misspelled word - doesn't seem to work though
}
\examples{
didYouMean('arse')
}

