% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{fetch}
\alias{fetch}
\title{Calculate wave fetch}
\usage{
fetch(x, y, rad, dir = 32, coast)
}
\arguments{
\item{x}{coord of point at wich to calc fetch.}

\item{y}{coord of point at wich to calc fetch.}

\item{rad}{Radius of fetch distance.}

\item{dir}{Number of compass directions to calc fetch.}

\item{coast}{SpatialPolgons of coastline to use.}
}
\value{
The fetch distance at each compass point.
}
\description{
Calculate wave fetch
}
\examples{
fetch(0, 0,1000,32,coast)
}

