% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{filter.by.dist}
\alias{filter.by.dist}
\title{filter.by.dist}
\usage{
filter.by.dist(data1, data2, datefield, d)
}
\arguments{
\item{data1}{}

\item{data2}{}

\item{datefield}{}

\item{d}{}
}
\description{
filter.by.dist
}

