% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomFuns.R
\name{folderStructure}
\alias{folderStructure}
\title{folderStructure}
\usage{
folderStructure(fld = getwd())
}
\arguments{
\item{fld}{Directory}
}
\description{
Create basic folder structure
}
\examples{
folderStructure()
}

