% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomFuns.R
\name{imageFF}
\alias{imageFF}
\title{imageFF
Apply a 'pie slice' filter to fourier transform of image}
\usage{
imageFF(r, d, l, t, lwl_buf = F, bd = 0.1)
}
\arguments{
\item{r}{raster}

\item{d}{direction}

\item{l}{radius}

\item{t}{taper}

\item{lwl_buf}{internal buffer radius}

\item{bd}{buffer width}
}
\description{
imageFF
Apply a 'pie slice' filter to fourier transform of image
}

