% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomFuns.R
\name{margin}
\alias{margin}
\title{margin
Returns the decision margin from a majority vote calssifier}
\usage{
margin(mat)
}
\arguments{
\item{mat}{matrix of votes from classifier}
}
\description{
margin
Returns the decision margin from a majority vote calssifier
}

