% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{owin2Polygons}
\alias{owin2Polygons}
\title{owin2Polygons
convert spatstat objects to sp classes}
\usage{
owin2Polygons(x, id = "1")
}
\arguments{
\item{x}{}

\item{id}{}
}
\description{
owin2Polygons
convert spatstat objects to sp classes
}

