% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomFuns.R
\name{partialPlotBi}
\alias{partialPlotBi}
\title{partialPlotBi}
\usage{
partialPlotBi(rf, y, x, v2, class, n.pt = 5)
}
\arguments{
\item{rf}{}

\item{y}{}

\item{x}{}

\item{v2}{}

\item{class}{}

\item{n.pt}{}
}
\description{
partialPlotBi
}

