% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomFuns.R
\name{permuteJB}
\alias{permuteJB}
\title{permuteJB}
\usage{
permuteJB(v1, v2, nreps = 10000)
}
\arguments{
\item{v1}{numeric vector}

\item{v2}{numeric vector}

\item{nreps}{number of imputations}
}
\value{
p-value
}
\description{
Jon Barrys original function for non-parametric sig test between means
}
\examples{
permuteJB(rnorm(100,0,sd = 1),rnorm(100,1,sd = 1))
}

