% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomFuns.R
\name{plotFF}
\alias{plotFF}
\title{plotFF
plots the fourier transformation of an image}
\usage{
plotFF(r, returnFF = F)
}
\arguments{
\item{r}{raster or matrix}

\item{returnFF}{}
}
\description{
plotFF
plots the fourier transformation of an image
}

