% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomFuns.R
\name{PP}
\alias{PP}
\title{PP
given vectors of kf,chl and IR, calculates primary Pproduction}
\usage{
PP(IR, kd, chl)
}
\arguments{
\item{IR}{}

\item{kd}{}

\item{chl}{}
}
\description{
PP
given vectors of kf,chl and IR, calculates primary Pproduction
}
\seealso{
Cloern 1987
}

