\name{randomFuns-package}
\alias{randomFuns-package}
\alias{randomFuns}
\docType{package}
\title{
My utility functions
}
\description{
My utility functions
}
\details{
\tabular{ll}{
Package: \tab randomFuns\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-05\cr
License: \tab What license is it under?\cr
}
-
}
\author{
-
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
