% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{RGBimage.seg}
\alias{RGBimage.seg}
\title{RGBimage.seg}
\usage{
RGBimage.seg(r, g, b, factor = 25, output)
}
\arguments{
\item{r}{}

\item{g}{}

\item{b}{}

\item{factor}{}

\item{output}{}
}
\description{
RGBimage.seg
}

