% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{Spatial.Point.Grid}
\alias{Spatial.Point.Grid}
\title{Spatial.Point.Grid}
\usage{
Spatial.Point.Grid(xlim, ylim, xres, yres)
}
\arguments{
\item{xlim}{}

\item{ylim}{}

\item{xres}{}

\item{yres}{}
}
\description{
Spatial.Point.Grid
}

