% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{stack.kmeans}
\alias{stack.kmeans}
\title{stack.kmeans}
\usage{
stack.kmeans(stack, cols = NA, cents = 3:5, plot = F, scale = F)
}
\arguments{
\item{stack}{}

\item{cols}{}

\item{cents}{}

\item{plot}{}

\item{scale}{}
}
\description{
stack.kmeans
}

