% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{standardise.origins}
\alias{standardise.origins}
\title{standardise.origins}
\usage{
standardise.origins(rasters, method = "bilinear")
}
\arguments{
\item{rasters}{}

\item{method}{}
}
\description{
standardise.origins
}

