% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomFuns.R
\name{swap.quadrant}
\alias{swap.quadrant}
\title{swap.quadrant
inverts the four quadrants of a matrix}
\usage{
swap.quadrant(ft)
}
\arguments{
\item{ft}{matrix}
}
\description{
swap.quadrant
inverts the four quadrants of a matrix
}

